+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 * Add `rawConnection` getter to `ConnectionContext`, which is a `SQLiteConnection` instance from
   `androidx.sqlite` that can be used to step through statements in a custom way.
 
+## 1.5.1
+
+* Fix issue in legacy sync client where local writes made offline could have their upload delayed
+  until a keepalive event was received. This could also cause downloaded updates to be delayed even
+  further until all uploads were
+  completed.
+* [Internal] Update core extension to 0.4.5
+
 ## 1.5.0
 
 * Add `PowerSyncDatabase.getCrudTransactions()`, returning a flow of transactions. This is useful